+++ contrib/test/v4l2grab.c
@@ -30,7 +30,7 @@ struct buffer {
 	size_t length;
 };
 
-static void xioctl(int fh, int request, void *arg)
+static void xioctl(int fh, unsigned long int request, void *arg)
 {
 	int r;
